requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
requests==2.31.0
python-jose[cryptography]==3.3.0
redis==5.0.1
slowapi==0.1.9
pydantic==2.6.1
